feat: Add slash command to set user's class
This commit adds a new slash command named "classe" to set the class of the user in the Discord bot. The command takes a single argument "classe" which represents the user's class (e.g., INF1-A, INF2-B, INF3-C). If the provided class is not valid, the bot responds with an error message. Otherwise, it confirms the successful class update.